polish-bank-manager-bot/requirements.txt
anyio==4.4.0
APScheduler==3.6.3
certifi==2024.7.4
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
python-dotenv==1.0.1
python-telegram-bot==13.3
pytz==2024.1
requests==2.32.3
setuptools==71.1.0
six==1.16.0
sniffio==1.3.1
tornado==6.4.1
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.15
